规范操作提示

+++ src/main/java/com/meiyouframework/bigwhale/controller/admin/cluster/AdminClusterController.java
@@ -1,21 +1,20 @@
 package com.meiyouframework.bigwhale.controller.admin.cluster;
 
-import com.meiyouframework.bigwhale.common.Constant;
 import com.meiyouframework.bigwhale.common.pojo.Msg;
 import com.meiyouframework.bigwhale.dto.DtoCluster;
+import com.meiyouframework.bigwhale.entity.Agent;
 import com.meiyouframework.bigwhale.entity.Cluster;
-import com.meiyouframework.bigwhale.service.ClusterService;
-import com.meiyouframework.bigwhale.service.MonitorService;
-import com.meiyouframework.bigwhale.service.ScriptService;
-import com.meiyouframework.bigwhale.service.SchedulingService;
+import com.meiyouframework.bigwhale.entity.Script;
+import com.meiyouframework.bigwhale.service.*;
 import com.meiyouframework.bigwhale.controller.BaseController;
-import com.meiyouframework.bigwhale.util.SchedulerUtils;
 import com.meiyouframework.bigwhale.util.WebHdfsUtils;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 
 @RestController
@@ -25,12 +24,9 @@ public class AdminClusterController extends BaseController {
     @Autowired
     private ClusterService clusterService;
     @Autowired
+    private AgentService agentService;
+    @Autowired
     private ScriptService scriptService;
-    @Autowired
-    private MonitorService monitorService;
-    @Autowired
-    private SchedulingService schedulingService;
-
 
     @RequestMapping(value = "/getpage.api", method = RequestMethod.POST)
     public Page<Cluster> getPage(@RequestBody DtoCluster req) {
@@ -48,11 +44,6 @@ public class AdminClusterController extends BaseController {
             if (dbCluster != null) {
                 return failed( "集群重复");
             }
-        } else {
-            Cluster dbCluster = clusterService.findById(req.getId());
-            if (dbCluster == null) {
-                return failed();
-            }
         }
         //创建目录
         boolean success = WebHdfsUtils.mkdir(req.getFsWebhdfs(), req.getFsDir(), req.getFsUser(), req.getFsUser());
@@ -69,22 +60,14 @@ public class AdminClusterController extends BaseController {
     public Msg delete(@RequestParam String id) {
         Cluster cluster = clusterService.findById(id);
         if (cluster != null) {
-            scriptService.findByQuery("clusterId=" + cluster.getId()).forEach(scriptInfo -> {
-                monitorService.findByQuery("scriptId=" + scriptInfo.getId()).forEach(item -> {
-                    try {
-                        SchedulerUtils.deleteJob(item.getId(), Constant.JobGroup.MONITOR);
-                    } catch (Exception e) {
-                        e.printStackTrace();
-                    }
-                });
-                schedulingService.findByQuery("scriptId=" + scriptInfo.getId()).forEach(item -> {
-                    try {
-                        SchedulerUtils.deleteJob(item.getId(), Constant.JobGroup.TIMED);
-                    } catch (Exception e) {
-                        e.printStackTrace();
-                    }
-                });
-            });
+            List<Script> scripts = scriptService.findByQuery("clusterId=" + cluster.getId());
+            if (!scripts.isEmpty()) {
+                return failed("集群下存在脚本，请先删除");
+            }
+            List<Agent> agents = agentService.findByQuery("clusterId=" + cluster.getId());
+            if (!agents.isEmpty()) {
+                return failed("集群下存在机器，请先删除");
+            }
             clusterService.delete(cluster);
         }
         return success();
